requirements.txt
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
DateTime==5.4
firebase-admin==6.3.0
future==0.18.3
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
iso8601==2.1.0
msgpack==1.0.7
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyserial==3.5
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
serial==0.0.97
uritemplate==4.1.1
urllib3==2.1.0
zope.interface==6.1
